collect openfermion obs in one file
